+++ src/altissimo/sendgrid/__init__.py
@@ -5,13 +5,13 @@ from __future__ import annotations
 from importlib.metadata import PackageNotFoundError, version
 
 from .client import EmailAddressLike, EmailRecipients, SendGridClient
-from .exceptions import SendGridError, SendGridImportError
+from .exceptions import SendGridError, SendGridImportError, SendGridSendError
 from .models import EmailAddress, SendResult
 
 try:
     __version__ = version("altissimo-sendgrid")
 except PackageNotFoundError:
-    __version__ = "0.1.0"
+    __version__ = "0.2.0"
 
 __all__ = [
     "EmailAddress",
@@ -20,6 +20,7 @@ __all__ = [
     "SendGridClient",
     "SendGridError",
     "SendGridImportError",
+    "SendGridSendError",
     "SendResult",
     "__version__",
 ]